requirements.txt
flask==3.1.0
pillow==10.4.0